* Added an experimental feature to allow commands in the test_manifest. So far, it's just "include", which grabs another file and starts reading test file names from that (recursively).
* This feature was suggested by Matt Vanderpol in January 2006, but
I'm just getting to it.

* For this release, I'm prefixing the command with a ";" as a way to
disambiguate the command name from a potential filename. I wanted to use
the "#" character to make it more C-like, but then I'd have to change
the comment character. I figure no one will have a ";" in their filename
(and now that I have typed that, so it has come to pass).


git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@1891 d9f60a02-fc28-0410-9752-bc0d6c4f57a6